pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.2",
    "inngest>=0.5.18",
    "llama-index-core>=0.14.19",
    "llama-index-readers-file>=0.6.0",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "requests>=2.32.0",
    "uvicorn>=0.42.0",
]
